Split the AXFRRetriever from resolver.cc
ALso, this moves the `parseResult` function into the pdns::resolver
namespace
