+++ pdns/recursordist/axfr-retriever.cc
@@ -1,0 +1,1 @@
+../axfr-retriever.cc
